src/main/kotlin/lab7/ShapeSerializer.kt
package lab7

import kotlinx.serialization.decodeFromString
import kotlinx.serialization.encodeToString
import kotlinx.serialization.json.Json
import kotlinx.serialization.modules.SerializersModule
import kotlinx.serialization.modules.polymorphic
import kotlinx.serialization.modules.subclass
import lab3.*

object ShapeSerializer {
    val json = Json {
        prettyPrint = true

        serializersModule = SerializersModule {
            polymorphic(Shape::class) {
                subclass(Square::class)
                subclass(Rectangle::class)
                subclass(Circle::class)
                subclass(Triangle::class)
            }
        }
    }

    inline fun <reified T> encode(value: T): String = json.encodeToString(value)

    inline fun <reified T> decode(string: String): T = json.decodeFromString(string)
}